# apps/api/openapi/school.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBodyDto": {
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "details": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "requestId"
        ],
        "type": "object"
      },
      "ApiErrorDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Capability": {
        "enum": [
          "school.settings.manage",
          "academic_year.manage",
          "class.manage",
          "section.manage",
          "subject.manage",
          "fee_head.manage",
          "holiday.manage",
          "user.account.manage",
          "role.manage",
          "student.view",
          "student.create",
          "student.update",
          "student.status.change",
          "enrolment.manage",
          "guardian.manage",
          "document.view",
          "document.upload",
          "document.verify",
          "staff.view",
          "staff.create",
          "staff.update",
          "staff.contract.manage",
          "staff.status.change",
          "staff.leave.approve",
          "payroll.view",
          "payroll.run",
          "attendance.student.mark",
          "attendance.student.view_all",
          "attendance.staff.manage",
          "assessment.define",
          "marks.enter",
          "marks.view_all",
          "result.approve",
          "result.publish",
          "diary.write",
          "remark.write",
          "timetable.manage",
          "certificate.issue",
          "charge.create",
          "charge.campaign.send",
          "concession.grant",
          "payment.record",
          "payment.verify",
          "payment.void",
          "collection.handover.confirm",
          "expense.record",
          "expense.approve",
          "finance.report.view",
          "fee.statement.view",
          "announcement.send.scope",
          "announcement.send.school"
        ],
        "type": "string"
      },
      "ChangeEmailDto": {
        "properties": {
          "currentPassword": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "format": "email",
            "maxLength": 254,
            "minLength": 3,
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "email"
        ],
        "type": "object"
      },
      "ChangePasswordDto": {
        "properties": {
          "currentPassword": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "newPassword": {
            "format": "password",
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "MALFORMED_REQUEST",
          "AUTH_REQUIRED",
          "AUTH_FAILED",
          "PERMISSION_DENIED",
          "SCHOOL_SUSPENDED",
          "ORIGIN_REJECTED",
          "NOT_FOUND",
          "VALIDATION_FAILED",
          "REFERENCE_NOT_FOUND",
          "UNKNOWN_FIELD",
          "INVALID_VALUE",
          "PAYLOAD_TOO_LARGE",
          "UNSUPPORTED_MEDIA_TYPE",
          "RATE_LIMITED",
          "SERVICE_UNAVAILABLE",
          "INTERNAL_ERROR",
          "PASSWORD_CHANGE_REQUIRED",
          "TOTP_REQUIRED",
          "TOTP_NOT_ENROLLED",
          "TOTP_ALREADY_ENROLLED",
          "TOTP_INVALID",
          "CURRENT_PASSWORD_INCORRECT",
          "SCHOOL_SHORT_CODE_TAKEN",
          "SCHOOL_SHORT_CODE_IMMUTABLE",
          "SCHOOL_TERMINATED",
          "ILLEGAL_STATUS_TRANSITION",
          "CONCURRENT_UPDATE",
          "TOKEN_INVALID",
          "EMAIL_NOT_VERIFIED",
          "SELF_ACTION_FORBIDDEN",
          "LAST_PRINCIPAL",
          "IDENTITY_NUMBER_MISSING",
          "ACTIVE_PRINCIPAL_EXISTS",
          "ALREADY_PRINCIPAL",
          "STAFF_NOT_ACTIVE",
          "USER_DISABLED",
          "ACADEMIC_YEAR_NAME_TAKEN",
          "ACADEMIC_YEAR_CLOSED",
          "ACADEMIC_YEAR_HAS_ACTIVE_ENROLMENTS",
          "CLASS_NAME_TAKEN",
          "CLASS_YEAR_IMMUTABLE",
          "CLASS_ARCHIVED",
          "CLASS_HAS_ACTIVE_ENROLMENTS",
          "SECTION_NAME_TAKEN",
          "SECTION_ARCHIVED",
          "SECTION_IN_USE",
          "SUBJECT_NAME_TAKEN",
          "SUBJECT_CODE_TAKEN",
          "SUBJECT_ARCHIVED",
          "GUARDIAN_CNIC_EXISTS",
          "GUARDIAN_CNIC_LOCKED",
          "GUARDIAN_MERGED",
          "GUARDIAN_CNIC_MISSING",
          "GUARDIAN_NO_LOGIN_LINK",
          "GUARDIAN_IS_PRIMARY_CONTACT",
          "LOGIN_ALREADY_EXISTS",
          "LINK_EXISTING_LOGIN_UNCONFIRMED",
          "UNEXPECTED_RESPONSE"
        ],
        "type": "string"
      },
...
      "MeDto": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "type": "array"
          },
          "email": {
            "nullable": true,
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "hasVerifiedEmail": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "passwordIsDefault": {
            "type": "boolean"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/SchoolRole"
            },
            "type": "array"
          },
          "school": {
            "$ref": "#/components/schemas/MeSchoolDto"
          },
          "sessionExpiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fullName",
          "email",
          "hasVerifiedEmail",
          "passwordIsDefault",
          "school",
          "roles",
          "capabilities",
          "sessionExpiresAt"
        ],
        "type": "object"
      },
      "MeSchoolDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortCode": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SchoolStatus"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "shortCode",
          "status"
        ],
        "type": "object"
      },
...
      "SchoolRole": {
        "enum": [
          "principal",
          "office_staff",
          "teacher",
          "parent",
          "student"
        ],
        "type": "string"
      },
      "SchoolSettingsDto": {
        "properties": {
          "feeDueDay": {
            "maximum": 28,
            "minimum": 1,
            "type": "number"
          },
          "studentLoginEnabled": {
            "type": "boolean"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "feeDueDay",
          "studentLoginEnabled",
          "updatedAt"
        ],
        "type": "object"
      },
      "SchoolStatus": {
        "enum": [
          "trial",
          "active",
          "suspended",
          "terminated"
        ],
        "type": "string"
      },
...
      "UpdateSchoolSettingsDto": {
        "properties": {
          "feeDueDay": {
            "maximum": 28,
            "minimum": 1,
            "type": "number"
          },
          "studentLoginEnabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "ASMS school API",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "MeController_get",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/change-email": {
      "post": {
        "operationId": "MeController_changeEmail",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/change-password": {
      "post": {
        "operationId": "MeController_changePassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/school/settings": {
      "get": {
        "operationId": "SchoolSettingsController_get",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolSettingsDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "school settings"
        ]
      },
      "patch": {
        "operationId": "SchoolSettingsController_update",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchoolSettingsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolSettingsDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "school settings"
        ]
      }
    },
...
  },
  "servers": [],
...
}